Add log correlation for log4j and log4j2
closes #922
