+++ apm-agent-benchmarks/pom.xml
@@ -47,7 +47,7 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>apm-slf4j-plugin</artifactId>
+            <artifactId>apm-log-correlation-plugin</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
