+++ apm-agent-plugins/apm-log-correlation-plugin/pom.xml
@@ -8,7 +8,7 @@
         <version>1.12.1-SNAPSHOT</version>
     </parent>
 
-    <artifactId>apm-slf4j-plugin</artifactId>
+    <artifactId>apm-log-correlation-plugin</artifactId>
     <name>${project.groupId}:${project.artifactId}</name>
 
     <properties>
@@ -21,6 +21,18 @@
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-classic</artifactId>
             <version>1.2.3</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.logging.log4j</groupId>
+            <artifactId>log4j-core</artifactId>
+            <version>2.12.0</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>log4j</groupId>
+            <artifactId>log4j</artifactId>
+            <version>1.2.17</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
