+++ apm-agent-plugins/apm-log-correlation-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.impl.ActivationListener
