+++ apm-agent-plugins/apm-log-correlation-plugin/src/test/java/co/elastic/apm/agent/slf4j/Slf4JMdcActivationListenerTest.java
@@ -11,9 +11,9 @@
  * the Apache License, Version 2.0 (the "License"); you may
  * not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -33,12 +33,16 @@ import co.elastic.apm.agent.impl.transaction.Span;
 import co.elastic.apm.agent.impl.transaction.TraceContext;
 import co.elastic.apm.agent.impl.transaction.Transaction;
 import co.elastic.apm.agent.logging.LoggingConfiguration;
+import org.apache.logging.log4j.ThreadContext;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.slf4j.MDC;
 
 import java.io.IOException;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -49,12 +53,25 @@ class Slf4JMdcActivationListenerTest extends AbstractInstrumentationTest {
 
     private LoggingConfiguration loggingConfiguration;
     private CoreConfiguration coreConfiguration;
+    private Boolean log4jMdcWorking;
+    private ExecutorService executorService = Executors.newSingleThreadExecutor();
 
     @BeforeEach
-    void setUp() {
+    void setUp() throws Exception {
+        org.apache.log4j.MDC.put("test", true);
+        log4jMdcWorking = (Boolean) org.apache.log4j.MDC.get("test");
         MDC.clear();
+        org.apache.log4j.MDC.clear();
+        ThreadContext.clearAll();
         loggingConfiguration = config.getConfig(LoggingConfiguration.class);
         coreConfiguration = config.getConfig(CoreConfiguration.class);
+        // initializes thread eagerly to avoid InheritableThreadLocal to inherit values to this thread
+        executorService.submit(() -> {}).get();
+    }
+
+    @AfterEach
+    void tearDown() {
+        executorService.shutdown();
     }
 
     @Test
@@ -234,17 +251,13 @@ class Slf4JMdcActivationListenerTest extends AbstractInstrumentationTest {
             final Span child = transaction.createSpan();
             try (Scope childScope = child.activateInScope()) {
                 assertMdcIsSet(child);
-                final CompletableFuture<Boolean> result = new CompletableFuture<>();
-                Thread thread = new Thread(() -> {
+                executorService.submit(() -> {
                     assertMdcIsEmpty();
                     try (Scope otherThreadScope = child.getTraceContext().activateInScope()) {
                         assertMdcIsSet(child);
                     }
                     assertMdcIsEmpty();
-                    result.complete(true);
-                });
-                thread.start();
-                assertThat(result.get(1000, TimeUnit.MILLISECONDS).booleanValue()).isTrue();
+                }).get();
                 assertMdcIsSet(child);
             }
             assertMdcIsSet(transaction);
@@ -267,11 +280,23 @@ class Slf4JMdcActivationListenerTest extends AbstractInstrumentationTest {
     private void assertMdcIsSet(AbstractSpan span) {
         assertThat(MDC.get("trace.id")).isEqualTo(span.getTraceContext().getTraceId().toString());
         assertThat(MDC.get("transaction.id")).isEqualTo(span.getTraceContext().getTransactionId().toString());
+        assertThat(ThreadContext.get("trace.id")).isEqualTo(span.getTraceContext().getTraceId().toString());
+        assertThat(ThreadContext.get("transaction.id")).isEqualTo(span.getTraceContext().getTransactionId().toString());
+        if (log4jMdcWorking == Boolean.TRUE) {
+            assertThat(org.apache.log4j.MDC.get("trace.id")).isEqualTo(span.getTraceContext().getTraceId().toString());
+            assertThat(org.apache.log4j.MDC.get("transaction.id")).isEqualTo(span.getTraceContext().getTransactionId().toString());
+        }
     }
 
     private void assertMdcIsEmpty() {
         assertThat(MDC.get("trace.id")).isNull();
         assertThat(MDC.get("transaction.id")).isNull();
+        assertThat(ThreadContext.get("trace.id")).isNull();
+        assertThat(ThreadContext.get("transaction.id")).isNull();
+        if (log4jMdcWorking == Boolean.TRUE) {
+            assertThat(org.apache.log4j.MDC.get("transaction.id")).isNull();
+            assertThat(org.apache.log4j.MDC.get("trace.id")).isNull();
+        }
     }
 
 }
