+++ apm-agent-plugins/pom.xml
@@ -28,7 +28,7 @@
         <module>apm-apache-httpclient-plugin</module>
         <module>apm-spring-resttemplate-plugin</module>
         <module>apm-httpclient-core</module>
-        <module>apm-slf4j-plugin</module>
+        <module>apm-log-correlation-plugin</module>
         <module>apm-es-restclient-plugin</module>
         <module>apm-okhttp-plugin</module>
         <module>apm-java-concurrent-plugin</module>
