+++ docs/supported-technologies.asciidoc
@@ -340,10 +340,23 @@ NOTE: only classes from packages configured in <<config-application-packages>>  
 
  Automatically <<api-span-capture-exception,captures exceptions>> for `logger.error("message", exception)` calls (since 1.10.0)
 |1.0.0
-|log4j
+
+|log4j2
 |2.0+
-|Automatically <<api-span-capture-exception,captures exceptions>> for `logger.error("message", exception)` calls.
-|1.11.0
+|When <<config-enable-log-correlation>> is set to `true`,
+the agent will add a https://logging.apache.org/log4j/2.x/manual/thread-context.html[ThreadContext] entry for `trace.id` and `transaction.id`.
+See the <<config-enable-log-correlation>> configuration option for more details.
+
+ Automatically <<api-span-capture-exception,captures exceptions>> for `logger.error("message", exception)` calls.
+|1.11.0 (error capturing), 1.13.0 (log correlation via `ThreadContext`)
+
+
+|log4j
+|1.x
+|When <<config-enable-log-correlation>> is set to `true`,
+the agent will add a https://logging.apache.org/log4j/1.2/apidocs/org/apache/log4j/MDC.html[MDC] entry for `trace.id` and `transaction.id`.
+See the <<config-enable-log-correlation>> configuration option for more details.
+|1.13.0
 
 |===
