+++ elastic-apm-agent/pom.xml
@@ -161,7 +161,7 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>apm-slf4j-plugin</artifactId>
+            <artifactId>apm-log-correlation-plugin</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
